requirements.txt
python-telegram-bot[job-queue]==20.7
openai>=1.30.0
python-dotenv>=1.0.0
